requirements.txt
mysql-connector-python==8.2.0
Pillow==10.1.0
protobuf==4.21.12
six==1.16.0
wxPython==4.2.1
